Added insert and update queries

+++ db/schema.sql
@@ -11,7 +11,7 @@ CREATE TABLE departments (
 CREATE TABLE roles (
   id INTEGER AUTO_INCREMENT PRIMARY KEY,
   title VARCHAR(30) NOT NULL,
-  salary DECIMAL (10, 2) NOT NULL,
+  salary INTEGER  NOT NULL,
   department_id INTEGER NOT NULL,
   CONSTRAINT fk_departments FOREIGN KEY (department_id) REFERENCES departments(id) ON DELETE CASCADE
 );
